Moved webapi to Windows 11 friendly name
