+++ Invoice_Gen.WebApi.sln
@@ -1,0 +1,27 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{E0C658C8-F6D9-4991-A98C-B6EF2400A08A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Invoice_Gen.WebApi", "src\Invoice_Gen.WebApi\Invoice_Gen.WebApi.csproj", "{F9489BCA-CA8F-4A00-9E5A-094204898D50}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F9489BCA-CA8F-4A00-9E5A-094204898D50}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F9489BCA-CA8F-4A00-9E5A-094204898D50}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F9489BCA-CA8F-4A00-9E5A-094204898D50}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F9489BCA-CA8F-4A00-9E5A-094204898D50}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{F9489BCA-CA8F-4A00-9E5A-094204898D50} = {E0C658C8-F6D9-4991-A98C-B6EF2400A08A}
+	EndGlobalSection
+EndGlobal
