--- src/App.WebApi/App.WebApi.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <VersionPrefix>0.0.0.1-beta</VersionPrefix>
-    <AssemblyName>App.WebApi</AssemblyName>
-    <PackageId>App.WebApi</PackageId>
-    <GenerateDocumentationFile>True</GenerateDocumentationFile>
-	  <NoWarn>$(NoWarn);1591</NoWarn>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Bogus" Version="34.0.2" />
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.5" />
-    <PackageReference Include="OwaspHeaders.Core" Version="6.0.5" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
-    <PackageReference Include="ClacksMiddlware" Version="2.1.0" />
-  </ItemGroup>
-
-</Project>
