--- src/App.WebApi/Models/HelloResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace App.WebApi.Models;
-
-public class HelloResponse
-{
-    public DateOnly Date { get; init; } = DateOnly.FromDateTime(DateTime.Now);
-    public string Message { get; init; } = "Hello there!";
-}
