--- src/App.WebApi/Services/IClientService.cs
@@ -1,8 +1,0 @@
-﻿using App.WebApi.Models;
-
-namespace App.WebApi.Services;
-
-public interface IClientService
-{
-    List<Client> GetClients();
-}
