+++ src/Invoice_Gen.WebApi/Controllers/VersionController.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
-using App.WebApi.Models;
+using Invoice_Gen.WebApi.Models;
 
-namespace App.WebApi.Controllers;
+namespace Invoice_Gen.WebApi.Controllers;
 
 // This controller currently responds to /
 // This is a quick way to find out if the app is running correctly.
@@ -11,10 +11,10 @@ namespace App.WebApi.Controllers;
 public class VersionController : ControllerBase
 {
     /// <summary>
-    /// Gets basic information about this app, wrapped in an instance of <see cref="App.WebApi.Models.VersionResponse" />
+    /// Gets basic information about this app, wrapped in an instance of <see cref="Invoice_Gen.WebApi.Models.VersionResponse" />
     /// </summary>
     /// <returns>
-    /// An instance of <see cref="App.WebApi.Models.VersionResponse" /> which represents the name and version number
+    /// An instance of <see cref="Invoice_Gen.WebApi.Models.VersionResponse" /> which represents the name and version number
     /// for the running application
     /// </returns>
     [HttpGet]
