+++ src/Invoice_Gen.WebApi/Models/HelloResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace Invoice_Gen.WebApi.Models;
+
+public class HelloResponse
+{
+    public DateOnly Date { get; init; } = DateOnly.FromDateTime(DateTime.Now);
+    public string Message { get; init; } = "Hello there!";
+}
