+++ src/Invoice_Gen.WebApi/Services/IClientService.cs
@@ -1,0 +1,8 @@
+﻿using Invoice_Gen.WebApi.Models;
+
+namespace Invoice_Gen.WebApi.Services;
+
+public interface IClientService
+{
+    List<Client> GetClients();
+}
